dbt/requirements-test.txt
# Testing Requirements for dbt Project
# Install with: pip install -r requirements-test.txt

# Core testing framework
pytest>=7.4.0
pytest-xdist>=3.3.0          # Parallel test execution
pytest-cov>=4.1.0            # Coverage reporting
pytest-html>=3.2.0           # HTML test reports
pytest-json-report>=1.5.0    # JSON test reports

# dbt testing utilities
dbt-core>=1.6.0
dbt-snowflake>=1.6.0
dbt-tests-adapter>=1.6.0

# Data validation and quality
great-expectations>=0.17.0    # Advanced data validation
pandera>=0.17.0              # DataFrame validation
cerberus>=1.3.0              # Schema validation

# Utilities for test development
pyyaml>=6.0.1                # YAML parsing
jinja2>=3.1.0                # Template testing
jsonschema>=4.19.0           # JSON schema validation
faker>=19.6.0                # Test data generation
factory-boy>=3.3.0           # Test object factories

# Performance and profiling
memory-profiler>=0.61.0      # Memory usage profiling
line-profiler>=4.1.0         # Line-by-line profiling
pytest-benchmark>=4.0.0     # Performance benchmarking

# Reporting and analysis
pandas>=2.0.0                # Data analysis
numpy>=1.24.0                # Numerical operations
matplotlib>=3.7.0            # Plotting for test reports
seaborn>=0.12.0              # Statistical plotting

# Documentation testing
sphinx>=7.1.0                # Documentation generation
sphinx-rtd-theme>=1.3.0      # Read the Docs theme
myst-parser>=2.0.0           # Markdown parsing for docs

# Code quality and linting
flake8>=6.0.0                # Code linting
black>=23.7.0                # Code formatting
isort>=5.12.0                # Import sorting
mypy>=1.5.0                  # Type checking

# Database connectivity for testing
snowflake-connector-python>=3.2.0
psycopg2-binary>=2.9.0       # PostgreSQL (if needed for testing)
sqlalchemy>=2.0.0            # Database abstraction

# Mocking and fixtures
responses>=0.23.0            # HTTP request mocking
freezegun>=1.2.0             # Time mocking
pytest-mock>=3.11.0         # Advanced mocking

# Parallel and distributed testing
pytest-xdist[psutil]>=3.3.0 # Distributed testing
pytest-parallel>=0.1.1      # Parallel test execution

# Test data and fixtures
mimesis>=11.1.0              # Test data generation
hypothesis>=6.82.0           # Property-based testing

# Environment and configuration
python-dotenv>=1.0.0         # Environment variable management
click>=8.1.0                 # CLI utilities
rich>=13.5.0                 # Rich terminal output
tabulate>=0.9.0              # Table formatting

# Monitoring and observability
prometheus-client>=0.17.0    # Metrics collection
structlog>=23.1.0            # Structured logging
